requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
bs4==0.0.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.6
django-background-tasks==1.2.5
django-compat==1.0.15
django-filter==24.2
django-oauth-toolkit==2.4.0
django-picklefield==3.2
django-q==1.3.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
environs==11.0.0
feedparser==6.0.11
idna==3.7
inflection==0.5.1
jinxed==1.2.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
Khayyam==3.0.17
marshmallow==3.21.2
oauthlib==3.2.2
packaging==24.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==3.5.3
referencing==0.35.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
rpds-py==0.18.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
